migrate rn functions from sdk-apple to sdk-rn
