+++ packages/sdk-apple/src/runner.ts
@@ -1,15 +1,12 @@
 import child_process, { ExecFileOptions } from 'child_process';
-import crypto from 'crypto';
 import path from 'path';
 import {
     fsExistsSync,
     mkdirSync,
     writeFileSync,
     fsWriteFileSync,
-    fsReadFileSync,
     copyFileSync,
     executeAsync,
-    commandExistsSync,
     getAppFolder,
     getConfigProp,
     isPlatformActive,
@@ -20,7 +17,6 @@ import {
     MACOS,
     TVOS,
     chalk,
-    logInfo,
     logTask,
     logError,
     logWarning,
@@ -41,130 +37,11 @@ import { parseXcodeProject } from './xcodeParser';
 import { parseXcscheme } from './xcschemeParser';
 import { AppleDevice, Context } from './types';
 import { ObjectEncodingOptions } from 'fs';
-import { packageReactNativeIOS, runReactNativeIOS } from '@rnv/sdk-react-native';
+import { packageReactNativeIOS, runCocoaPods, runReactNativeIOS } from '@rnv/sdk-react-native';
 import { registerDevice } from './fastlane';
-import { EnvVars } from './env';
-
-const printableEnvKeys = [
-    'RCT_NEW_ARCH_ENABLED',
-    'REACT_NATIVE_PERMISSIONS_REQUIRED',
-    'RNV_APP_BUILD_DIR',
-    'RNV_ENGINE_PATH',
-];
 
 export const packageBundleForXcode = (c: Context) => {
     return packageReactNativeIOS(c);
-};
-
-export const generateChecksum = (str: string, algorithm?: string, encoding?: 'base64' | 'base64url' | 'hex') =>
-    crypto
-        .createHash(algorithm || 'md5')
-        .update(str, 'utf8')
-        .digest(encoding || 'hex');
-
-const checkIfPodsIsRequired = async (c: Context) => {
-    const appFolder = getAppFolder(c);
-    const podChecksumPath = path.join(appFolder, 'Podfile.checksum');
-    if (!fsExistsSync(podChecksumPath)) return true;
-    const podChecksum = fsReadFileSync(podChecksumPath).toString();
-    const podContentChecksum = generateChecksum(fsReadFileSync(path.join(appFolder, 'Podfile')).toString());
-    const packageDependenciesChecksum = generateChecksum(
-        JSON.stringify({ ...c.files.project.package.dependencies, ...c.files.project.package.devDependencies })
-    );
-    const combinedChecksum = podContentChecksum + packageDependenciesChecksum;
-
-    if (podChecksum !== combinedChecksum) {
-        logDebug('runCocoaPods:isMandatory');
-        return true;
-    }
-    logInfo(
-        'Pods do not seem like they need to be updated. If you want to update them manually run the same command with "-u" parameter'
-    );
-    return false;
-};
-
-const updatePodsChecksum = (c: Context) => {
-    logTask('updatePodsChecksum');
-    const appFolder = getAppFolder(c);
-    const podChecksumPath = path.join(appFolder, 'Podfile.checksum');
-    const podContentChecksum = generateChecksum(fsReadFileSync(path.join(appFolder, 'Podfile')).toString());
-    const packageDependenciesChecksum = generateChecksum(
-        JSON.stringify({ ...c.files.project.package.dependencies, ...c.files.project.package.devDependencies })
-    );
-    const combinedChecksum = podContentChecksum + packageDependenciesChecksum;
-    if (fsExistsSync(podChecksumPath)) {
-        const existingContent = fsReadFileSync(podChecksumPath).toString();
-        if (existingContent !== combinedChecksum) {
-            logDebug(`updatePodsChecksum:${combinedChecksum}`);
-            return fsWriteFileSync(podChecksumPath, combinedChecksum);
-        }
-        return true;
-    }
-    logDebug(`updatePodsChecksum:${combinedChecksum}`);
-    return fsWriteFileSync(podChecksumPath, combinedChecksum);
-};
-
-const runCocoaPods = async (c: Context) => {
-    logTask('runCocoaPods', `forceUpdate:${!!c.program.updatePods}`);
-
-    if (c.runtime._skipNativeDepResolutions) return;
-
-    const appFolder = getAppFolder(c);
-
-    if (!fsExistsSync(appFolder)) {
-        return Promise.reject(`Location ${appFolder} does not exists!`);
-    }
-    const podsRequired = c.program.updatePods || (await checkIfPodsIsRequired(c));
-
-    const env: any = {
-        ...process.env,
-        ...CoreEnvVars.BASE(),
-        ...EnvVars.REACT_NATIVE_PERMISSIONS_REQUIRED(),
-        ...EnvVars.RCT_NEW_ARCH_ENABLED(),
-    };
-
-    if (podsRequired) {
-        if (!commandExistsSync('pod')) {
-            throw new Error('Cocoapods not installed. Please run `sudo gem install cocoapods`');
-        }
-
-        try {
-            await executeAsync(c, 'bundle install', {
-                env: process.env,
-                printableEnvKeys,
-            });
-            await executeAsync(c, 'bundle exec pod install', {
-                cwd: appFolder,
-                env,
-                printableEnvKeys,
-            });
-        } catch (e: Error | any) {
-            const s = e?.toString ? e.toString() : '';
-            const isGenericError =
-                s.includes('No provisionProfileSpecifier configured') ||
-                s.includes('TypeError:') ||
-                s.includes('ReferenceError:') ||
-                s.includes('find gem cocoapods');
-            if (isGenericError) {
-                return new Error(`pod install failed with:\n ${s}`);
-            }
-            logWarning(`pod install is not enough! Let's try pod update! Error:\n ${s}`);
-            await executeAsync(c, 'bundle update', {
-                env: process.env,
-            });
-
-            return executeAsync(c, 'bundle exec pod update', {
-                cwd: appFolder,
-                env,
-                printableEnvKeys,
-            })
-                .then(() => updatePodsChecksum(c))
-                .catch((er) => Promise.reject(er));
-        }
-
-        updatePodsChecksum(c);
-        return true;
-    }
 };
 
 const copyAppleAssets = (c: Context, platform: RnvPlatform, appFolderName: string) =>
