+++ packages/sdk-react-native/src/iosRunner.ts
@@ -1,7 +1,24 @@
-import { executeAsync, getAppFolder, getConfigProp, doResolve, logTask, RnvContext, CoreEnvVars } from '@rnv/core';
+import {
+    executeAsync,
+    getAppFolder,
+    getConfigProp,
+    doResolve,
+    logTask,
+    RnvContext,
+    CoreEnvVars,
+    fsReadFileSync,
+    fsExistsSync,
+    logInfo,
+    logDebug,
+    fsWriteFileSync,
+    commandExistsSync,
+    logWarning,
+} from '@rnv/core';
 import { RnvEnvContext } from '@rnv/core/lib/env/types';
 import { EnvVars, printableEnvKeys } from './env';
 import shellQuote from 'shell-quote';
+import path from 'path';
+import crypto from 'crypto';
 
 export const packageReactNativeIOS = (c: RnvContext, isDev = false) => {
     logTask('packageBundleForXcode');
@@ -85,5 +102,116 @@ export const runReactNativeIOS = async (
         } else if (e instanceof Error) {
             return Promise.reject(e.message);
         }
+    }
+};
+export const generateChecksum = (str: string, algorithm?: string, encoding?: 'base64' | 'base64url' | 'hex') =>
+    crypto
+        .createHash(algorithm || 'md5')
+        .update(str, 'utf8')
+        .digest(encoding || 'hex');
+
+const checkIfPodsIsRequired = async (c: RnvContext) => {
+    const appFolder = getAppFolder(c);
+    const podChecksumPath = path.join(appFolder, 'Podfile.checksum');
+    if (!fsExistsSync(podChecksumPath)) return true;
+    const podChecksum = fsReadFileSync(podChecksumPath).toString();
+    const podContentChecksum = generateChecksum(fsReadFileSync(path.join(appFolder, 'Podfile')).toString());
+    const packageDependenciesChecksum = generateChecksum(
+        JSON.stringify({ ...c.files.project.package.dependencies, ...c.files.project.package.devDependencies })
+    );
+    const combinedChecksum = podContentChecksum + packageDependenciesChecksum;
+
+    if (podChecksum !== combinedChecksum) {
+        logDebug('runCocoaPods:isMandatory');
+        return true;
+    }
+    logInfo(
+        'Pods do not seem like they need to be updated. If you want to update them manually run the same command with "-u" parameter'
+    );
+    return false;
+};
+
+const updatePodsChecksum = (c: RnvContext) => {
+    logTask('updatePodsChecksum');
+    const appFolder = getAppFolder(c);
+    const podChecksumPath = path.join(appFolder, 'Podfile.checksum');
+    const podContentChecksum = generateChecksum(fsReadFileSync(path.join(appFolder, 'Podfile')).toString());
+    const packageDependenciesChecksum = generateChecksum(
+        JSON.stringify({ ...c.files.project.package.dependencies, ...c.files.project.package.devDependencies })
+    );
+    const combinedChecksum = podContentChecksum + packageDependenciesChecksum;
+    if (fsExistsSync(podChecksumPath)) {
+        const existingContent = fsReadFileSync(podChecksumPath).toString();
+        if (existingContent !== combinedChecksum) {
+            logDebug(`updatePodsChecksum:${combinedChecksum}`);
+            return fsWriteFileSync(podChecksumPath, combinedChecksum);
+        }
+        return true;
+    }
+    logDebug(`updatePodsChecksum:${combinedChecksum}`);
+    return fsWriteFileSync(podChecksumPath, combinedChecksum);
+};
+
+export const runCocoaPods = async (c: RnvContext) => {
+    logTask('runCocoaPods', `forceUpdate:${!!c.program.updatePods}`);
+
+    if (c.runtime._skipNativeDepResolutions) return;
+
+    const appFolder = getAppFolder(c);
+
+    if (!fsExistsSync(appFolder)) {
+        return Promise.reject(`Location ${appFolder} does not exists!`);
+    }
+    const podsRequired = c.program.updatePods || (await checkIfPodsIsRequired(c));
+
+    const env: any = {
+        ...process.env,
+        ...CoreEnvVars.BASE(),
+        ...EnvVars.RNV_REACT_NATIVE_PATH(),
+        ...EnvVars.REACT_NATIVE_PERMISSIONS_REQUIRED(),
+        ...EnvVars.RCT_NEW_ARCH_ENABLED(),
+    };
+
+    if (podsRequired) {
+        if (!commandExistsSync('pod')) {
+            throw new Error('Cocoapods not installed. Please run `sudo gem install cocoapods`');
+        }
+
+        try {
+            await executeAsync(c, 'bundle install', {
+                env: process.env,
+                printableEnvKeys,
+            });
+            await executeAsync(c, 'bundle exec pod install', {
+                cwd: appFolder,
+                env,
+                printableEnvKeys,
+            });
+        } catch (e: Error | any) {
+            const s = e?.toString ? e.toString() : '';
+            const isGenericError =
+                s.includes('No provisionProfileSpecifier configured') ||
+                s.includes('TypeError:') ||
+                s.includes('ReferenceError:') ||
+                s.includes('find gem cocoapods');
+            if (isGenericError) {
+                return new Error(`pod install failed with:\n ${s}`);
+            }
+            logWarning(`pod install is not enough! Let's try pod update! Error:\n ${s}`);
+            await executeAsync(c, 'bundle update', {
+                env: process.env,
+            });
+
+            return executeAsync(c, 'bundle exec pod update', {
+                cwd: appFolder,
+                env,
+                printableEnvKeys,
+            })
+                .then(() => updatePodsChecksum(c))
+                .catch((er) => Promise.reject(er));
+        }
+
+        updatePodsChecksum(c);
+        return true;
     }
 };
