+++ packages/sdk-react-native/src/rnPermissionsMap.ts
